requirements.txt
flask==2.0.3
flask[async]==2.0.3
nest_asyncio==1.5.4
requests==2.27.1
pytiktokapi==0.0.5
gunicorn==20.1.0
